+++ restli-int-test/src/test/java/com/linkedin/restli/examples/TestParseqBasedFluentClientApiWithProjections.java
@@ -16,7 +16,6 @@
 
 package com.linkedin.restli.examples;
 
-import com.linkedin.data.schema.MaskMap;
 import com.linkedin.parseq.ParSeqUnitTestHelper;
 import com.linkedin.restli.client.ParSeqRestliClient;
 import com.linkedin.restli.client.ParSeqRestliClientBuilder;
@@ -100,9 +99,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     ManualProjectionsFluentClient projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     CompletionStage<Greeting> result = projections.get(1L, false,
-        new ManualProjectionsFluentClient.GetOptionalParameters().setFields(
-            Greeting.createMask().withMessage()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withMessage()));
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertFalse(greeting.hasId());
@@ -117,9 +114,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
 
     Set<Long> ids = Sets.newHashSet(Arrays.asList(1L, 2L, 3L));
     CompletionStage<Map<Long, EntityResponse<Greeting>>> result = greetings.batchGet(ids,
-        new GreetingsFluentClient.BatchGetOptionalParameters().setFields(
-            Greeting.createMask().withTone()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withTone()));
     CompletableFuture<Map<Long, EntityResponse<Greeting>>> future = result.toCompletableFuture();
     Map<Long, EntityResponse<Greeting>> resultMap = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertEquals(resultMap.size(), ids.size());
@@ -141,9 +136,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
 
     String msg = Double.toString(Math.random());
     CompletionStage<IdEntityResponse<Long, Greeting>> result = greetings.createAndGet(getGreeting(msg),
-        new CreateGreetingFluentClient.CreateOptionalParameters().setFields(
-            Greeting.createMask().withMessage()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withMessage()));
     CompletableFuture<IdEntityResponse<Long, Greeting>> future = result.toCompletableFuture();
     Assert.assertNotNull(future.get(5000, TimeUnit.MILLISECONDS));
     Assert.assertFalse(future.get().getEntity().hasId());
@@ -160,9 +153,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     String msg2 = Double.toString(Math.random());
     CompletionStage<List<CreateIdEntityStatus<Long, Greeting>>>
         result = greetings.batchCreateAndGet(Arrays.asList(getGreeting(msg1), getGreeting(msg2)),
-            new CreateGreetingFluentClient.BatchCreateOptionalParameters().setFields(
-                Greeting.createMask().withId()
-            ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withId()));
     CompletableFuture<List<CreateIdEntityStatus<Long, Greeting>>> future = result.toCompletableFuture();
     List<CreateIdEntityStatus<Long, Greeting>> entities = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertEquals(entities.size(), 2);
@@ -183,9 +174,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     String message = "Edited message: fluent api test partialUpdateAndGet";
     Greeting update = getGreeting(message);
     CompletionStage<Greeting> result = greetings.partialUpdateAndGet(21L, PatchGenerator.diff(original, update),
-        new PartialUpdateGreetingFluentClient.PartialUpdateOptionalParameters().setFields(
-            Greeting.createMask().withId()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withId()));
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertFalse(greeting.hasMessage());
@@ -204,9 +193,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     inputs.put(21L, PatchGenerator.diff(original, update));
     inputs.put(22L, PatchGenerator.diff(original, update));
     CompletionStage<Map<Long, UpdateEntityStatus<Greeting>>> result = greetings.batchPartialUpdateAndGet(inputs,
-        new PartialUpdateGreetingFluentClient.BatchPartialUpdateOptionalParameters().setFields(
-            Greeting.createMask().withId().withMessage().withTone()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withId().withMessage().withTone()));
     CompletableFuture<Map<Long, UpdateEntityStatus<Greeting>>> future = result.toCompletableFuture();
     Assert.assertNotNull(future.get(5000, TimeUnit.MILLISECONDS));
     Assert.assertEquals(future.get().get(21L).getEntity().getId().longValue(), 21L);
@@ -225,10 +212,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
         Arrays.asList(getGreeting("GetAll").setId(200L), getGreeting("GetAll").setId(201L)));
 
     CompletionStage<List<Greeting>> result = createResult.thenCompose(ids -> greetings.getAll(
-        new GreetingsFluentClient.GetAllOptionalParameters().setFields(
-            Greeting.createMask().withMessage()
-        )
-    ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withMessage())));
     CompletableFuture<List<Greeting>> future = result.toCompletableFuture();
     List<Greeting> greetingList = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertTrue(greetingList.size() >= 2);
@@ -260,9 +244,7 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
     ManualProjectionsFluentClient projections = new ManualProjectionsFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
 
     CompletionStage<Greeting> result = projections.get(1L, true,
-        new ManualProjectionsFluentClient.GetOptionalParameters().setFields(
-            Greeting.createMask().withMessage()
-        ));
+        optionalParams -> optionalParams.withFields(mask -> mask.withMessage()));
     CompletableFuture<Greeting> future = result.toCompletableFuture();
     Greeting greeting = future.get(5000, TimeUnit.MILLISECONDS);
     // these fields would have been excluded by the framework if automatic projection was enabled
@@ -292,16 +274,14 @@ public class TestParseqBasedFluentClientApiWithProjections extends RestLiIntegra
   public void testValidationWithOnlyValidFieldsProjected() throws Exception
   {
     AutoValidationWithProjectionFluentClient validationDemos = new AutoValidationWithProjectionFluentClient(_parSeqRestliClient, _parSeqUnitTestHelper.getEngine());
-    MaskMap mask = ValidationDemo.createMask()
-        .withStringB()
-        .withIncludedB()
-        .withUnionFieldWithInlineRecord(m1 -> m1.withMyRecord(myRecord.ProjectionMask::withFoo2))
-        .withArrayWithInlineRecord(itemMask -> itemMask.withItems(myItem.ProjectionMask::withBar1))
-        .withMapWithTyperefs(m -> m.withValues(Greeting.ProjectionMask::withId))
-        .withValidationDemoNext(ValidationDemo.ProjectionMask::withIntB);
 
     CompletionStage<ValidationDemo> result = validationDemos.get(1,
-        new AutoValidationWithProjectionFluentClient.GetOptionalParameters().setFields(mask));
+        optionalParams -> optionalParams.withFields(mask -> mask.withStringB()
+            .withIncludedB()
+            .withUnionFieldWithInlineRecord(m1 -> m1.withMyRecord(myRecord.ProjectionMask::withFoo2))
+            .withArrayWithInlineRecord(itemMask -> itemMask.withItems(myItem.ProjectionMask::withBar1))
+            .withMapWithTyperefs(m -> m.withValues(Greeting.ProjectionMask::withId))
+            .withValidationDemoNext(ValidationDemo.ProjectionMask::withIntB)));
     CompletableFuture<ValidationDemo> future = result.toCompletableFuture();
     ValidationDemo validationDemo = future.get(5000, TimeUnit.MILLISECONDS);
     Assert.assertNotNull(validationDemo);